Move UI components into the components folder
