+++ src/ui/components/drop_zone/styles.css
