+++ src/ui/components/root/index.tsx
@@ -1,10 +1,10 @@
 import { Show } from 'solid-js'
 import type { Component } from 'solid-js'
-import { useStore } from './app'
-import { DropZone } from './drop_zone'
-import { Editor } from './editor'
-import { Toolbar } from './toolbar/toolbar'
-import { Styles } from '/src/ui/styles'
+import { useStore } from '../../app'
+import { DropZone } from '../drop_zone'
+import { Editor } from '../editor'
+import { Toolbar } from '../toolbar'
+import { Styles } from './styles'
 import { override } from '/src/utils/merge'
 import type InkInternal from '/types/internal'
