+++ src/ui/components/root/styles.tsx
@@ -1,8 +1,8 @@
 import { type Component, createEffect, createSignal, onMount } from 'solid-js'
 import { buildVendorUpdates } from '/src/extensions'
+import { makeVars } from '../../utils'
+import { useStore } from '../../app'
 import styles from './styles.css?inline'
-import { makeVars } from './utils'
-import { useStore } from './app'
 
 export const Styles: Component = () => {
   const [state] = useStore()
